Rename Xgit.Repository.WorkingTree.WriteIndexFile.to_iodevice/1 to Xgit.DirCache.to_iodevice/1.
